// src/libs/twitter/types/tweet/tweet.ts
export interface Result {
    __typename: string;
    rest_id: string;
    core: Core;
    unmention_info: UnmentionInfo;
    legacy: Legacy2;
    quick_promote_eligibility: QuickPromoteEligibility;
}

export interface Core {
    user_results: UserResults;
}

export interface UserResults {
    result: Result2;
}
// ...
export interface UnmentionInfo {}

export interface Legacy2 {
    created_at: string;
    conversation_id_str: string;
    display_text_range: number[];
    entities: Entities2;
    extended_entities: _ExtendedEntities;
    favorite_count: number;
    favorited: boolean;
    full_text: string;
    in_reply_to_status_id_str: string;
    is_quote_status: boolean;
    lang: string;
    possibly_sensitive: boolean;
    possibly_sensitive_editable: boolean;
    quote_count: number;
    quoted_status_id_str: string;
    reply_count: number;
    retweet_count: number;
    retweeted: boolean;
    source: string;
    user_id_str: string;
    id_str: string;
    self_thread: SelfThread;
}

export interface Entities2 {
    media: Medum[];
    user_mentions: any[];
    urls: any[];
    hashtags: any[];
    symbols: any[];
}

export interface Medum {
    display_url: string;
    expanded_url: string;
    id_str: string;
    indices: number[];
    media_url_https: string;
    type: string;
    url: string;
    features: Features;
    sizes: Sizes;
    original_info: OriginalInfo;
}

export interface Features {
    large: Large;
    medium: Medium;
    small: Small;
    orig: Orig;
}

export interface Large {
    faces: any[];
}

export interface Medium {
    faces: any[];
}

export interface Small {
    faces: any[];
}

export interface Orig {
    faces: any[];
}

export interface Sizes {
    large: Large2;
    medium: Medium2;
    small: Small2;
    thumb: Thumb;
}

export interface Large2 {
    h: number;
    w: number;
    resize: string;
}

export interface Medium2 {
    h: number;
    w: number;
    resize: string;
}

export interface Small2 {
    h: number;
    w: number;
    resize: string;
}

export interface Thumb {
    h: number;
    w: number;
    resize: string;
}

export interface OriginalInfo {
    height: number;
    width: number;
    focus_rects: FocusRect[];
}
// ...
export interface SelfThread {
    id_str: string;
}

export interface QuickPromoteEligibility {
    eligibility: string;
}